avatar/pyproject.toml
[project]
name = "new-realtimeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.11.0.86",
    "pipecat-ai[google,runner,simli,webrtc,websocket]>=0.0.84",
]
